Move DTO models in their own files
